Module sarzak::drawing::types

Types for instances of the “Drawing” domain

The following types are defined herein:

Anchor

BinaryUi

Point

ObjectEdge

TOP

LEFT

Edge

RelationshipUi

ObjectUi

IsaUi

AssociativeUi

SubtypeAnchors

RIGHT

BOTTOM

Generated Code – do not edit – yet. Use the following invocation to reproduce:

 target/debug/sarzak --meta gen -p ../nut/crates/sarzak

Structs

Anchor

An anchor, or anchor point, is the location where an arrow from a relationship attached to an object.

AssociativeUi

BinaryUi

This represents additional information necessary to render a Binary relationship in the user interface.

IsaUi

This represents additional data necessary to render an Isa relationship in the user interface .

ObjectEdge

The Edge of an Object Depiction

ObjectUi

Render a rectangle

Point

A point is a two-tuple that represents a location on the drawing canvas.

SubtypeAnchors

Subtype Anchors

Enums

Edge

An attachment point for an Anchor

RelationshipUi

Additional information necessary to render relationships in the user interface.

Constants

BOTTOM

The bottom of a rendered box

LEFT

The left side of a rendered box

RIGHT

The right side of a rendered box

TOP

The top edge of the rendered box